feat: implement donation processing with improved feedback
- Integrated makeDonation action for handling donation submissions.
- Replaced alert messages with toast notifications for success and error feedback.
- Updated donation form and cards to streamline donation processing and user experience.
- Enhanced error handling during donation submission to provide clearer user guidance.
